Update tokenizer.py
Now it gets all the content from the html and filters out all punctuation and non-alphanumeric characters without getting hyphens so we get a list of all the words in the url. It also includes apostrophes surrounded by alphanumeric characters on both sides because the ones at the end aren't really "in" the words. This also makes it so that compute_word_frequencies can actually get the stop words in a reasonable way as opposed to "don" and "t", "won" and "t", etc.